Widen core/simulation precision to double; make dynamic-substep cap configurable
Follow-up to the dynamic-timestep code review: widens core halo-tracking
precision from float to double (uncovering and fixing a real bug in the
process), makes the dynamic-timestep substep ceiling a run-file parameter
instead of a hardcoded constant, and documents the sage_satellite_stripping
N-dependence as inherited SAGE behavior rather than a Mimic defect.

Core/simulation precision (Mimic is not sage16-only; core must be highest
quality for every model):
- src/core/core_properties.yaml: Mvir, deltaMvir, CentralMvir, Rvir, Vvir,
  infallMvir, infallVvir, infallVmax widened from float to double. These are
  shared by every model (sage16, sham, halos-only), so the prior float choice
  (inherited from sage-model's struct GALAXY) had no parity justification at
  the core level.
- This surfaced a genuine bug: apply_descendant_properties() in
  src/core/inheritance.c tracks each halo's historical-maximum Rvir/Vvir via
  `if (descendant->virial_mass > halo->Mvir)`, comparing a fresh double mass
  against the previous snapshot's stored mass. Under float storage that
  comparison could silently go the wrong way near the rounding boundary,
  freezing an orphan's preserved Rvir/Vvir at an earlier, smaller snapshot.
  Double storage makes the comparison exact.
- src/module_system/output_helpers.h: output_rvir_conditional,
  output_vvir_conditional, output_infall_property_or_zero were hardcoded to
  return/accept float with explicit casts, discarding the new precision;
  widened to double.
- src/core/build_model.c: a stray `float central_mvir` local (feeding
  CentralMvir) widened to double for the same reason.
- tests/data/output/baseline/{binary,hdf5}/*: regenerated via
  scripts/regenerate_baseline.sh (HDF5) and the documented manual copy
  (binary) after confirming all 43 unit tests, sampled integration tests, and
  the scientific virial-relation check pass. ~0.8-1% of mini-Millennium
  halos now show more accurate historical-maximum Rvir/Vvir.

Dynamic-timestep substep ceiling made configurable (mini-Millennium analysis
showed the old hardcoded 50 dominated ~73-84% of transitions at the
SubSteps values users would actually pick, and the right ceiling depends on
simulation snapshot spacing, not just SubSteps):
- src/include/constants.h: MAX_DYNAMIC_SUBSTEPS (50) replaced with
  DEFAULT_MAX_DYNAMIC_SUBSTEPS (200), now a default rather than a hard cap.
- src/include/types.h, src/include/proto.h, src/core/timestep.c,
  src/core/build_model.c, src/core/read_parameter_file.c: added optional
  MaxDynamicSubsteps run-file parameter, threaded through
  compute_dynamic_substeps(), parsed/validated/logged like TimestepScheme.
- src/io/output/metadata_hdf5.c: MaxDynamicSubsteps recorded in HDF5
  RunProperties alongside SubSteps/TimestepScheme.
- docs/USER-GUIDE.md, docs/DEVELOPER-GUIDE.md: documented the new parameter
  and its role as a safety ceiling, not a resolution target.
- tests/unit/test_dynamic_substeps.c, tests/unit/test_parameter_parsing.c:
  added parse/validation/formula coverage for the new parameter.
- tests/framework/harness.py, tests/integration/test_output_formats.py:
  added max_dynamic_substeps override plumbing and RunProperties round-trip
  coverage.

sage_satellite_stripping N-dependence (confirmed identical in stock
sage-model's model_infall.c, not a Mimic-introduced bug):
- models/sage16/modules/sage_satellite_stripping/sage_satellite_stripping.c:
  documented why the stripped fraction follows 1-(1-1/N)^N and why it is
  preserved for SAGE parity rather than fixed.
- models/sage16/modules/sage_satellite_stripping/_tests/..., .../sage_apply_infall/_tests/...:
  added N-invariance/characterization regression tests
  (test_stripped_fraction_follows_geometric_formula,
  test_physics_infall_budget_invariant_across_substep_counts), verified
  against the built module rather than only hand-derived.

Documentation:
- models/sage16/model_properties.yaml: noted the float gas/mass/metal
  reservoirs are a deliberate sage-model parity choice, not a default
  recommendation.
- docs/DEVELOPER-GUIDE.md: new Property Precision guidance for choosing
  float vs double in new models/simulations.
- docs/dev/DYNAMIC-TIMESTEP-CONVERGENCE-NOTES.md (new): working analysis
  notes from this review pass - substep-pattern taxonomy, the mini-Millennium
  MaxDynamicSubsteps sizing data, the core/simulation precision findings, and
  a sketch for a future N-invariant satellite-stripping module (not
  implemented here; sage_satellite_stripping itself is untouched
  behaviorally).

+++ src/module_system/output_helpers.h
@@ -46,8 +46,8 @@ struct Halo;
  *     output_function: output_infall_property_or_zero
  *     output_function_arg: "g, g->infallMvir"
  */
-static inline float output_infall_property_or_zero(const struct Halo *g, float value) {
-  return (g->Type != 0) ? value : 0.0f;
+static inline double output_infall_property_or_zero(const struct Halo *g, double value) {
+  return (g->Type != 0) ? value : 0.0;
 }
 
 /**
@@ -58,10 +58,10 @@ static inline float output_infall_property_or_zero(const struct Halo *g, float v
  *
  * Used by: Rvir property
  */
-static inline float output_rvir_conditional(const struct Halo *g) {
+static inline double output_rvir_conditional(const struct Halo *g) {
   /* Type 2 orphans: preserve value from when the orphan had a subhalo;
    * Type 0/1: recalculate current (stored value is "maximum ever") */
-  return (g->Type == 2) ? g->Rvir : (float)get_virial_radius(g->HaloNr);
+  return (g->Type == 2) ? g->Rvir : get_virial_radius(g->HaloNr);
 }
 
 /**
@@ -72,10 +72,10 @@ static inline float output_rvir_conditional(const struct Halo *g) {
  *
  * Used by: Vvir property
  */
-static inline float output_vvir_conditional(const struct Halo *g) {
+static inline double output_vvir_conditional(const struct Halo *g) {
   /* Type 2 orphans: preserve value from when the orphan had a subhalo;
    * Type 0/1: recalculate current (stored value is "maximum ever") */
-  return (g->Type == 2) ? g->Vvir : (float)get_virial_velocity(g->HaloNr);
+  return (g->Type == 2) ? g->Vvir : get_virial_velocity(g->HaloNr);
 }
 
 #endif /* OUTPUT_HELPERS_H */
